edenapi_server: Remove 'pretty' param from repos endpoint
Summary: The intention of this parameter was to produce human-readable JSON for tests and for manual debugging. In practice, given that the JSON response here is just a (typically short) list, there isn't really much value in prettifying. We weren't using this anyway, so             let's remove it to simplify the code.

Reviewed By: xavierd

Differential Revision: D22020784

fbshipit-source-id: 09290cc667cfdb194693e6c05858af2f8905f88e